+++ Plans/Sentry_TestServer.xctestplan
@@ -1,0 +1,24 @@
+{
+  "configurations" : [
+    {
+      "id" : "237352B3-25E4-4518-94D0-F0951A385694",
+      "name" : "Configuration 1",
+      "options" : {
+
+      }
+    }
+  ],
+  "defaultOptions" : {
+    "testTimeoutsEnabled" : true
+  },
+  "testTargets" : [
+    {
+      "target" : {
+        "containerPath" : "container:Sentry.xcodeproj",
+        "identifier" : "62FE9E292E7996CD00A6BD0D",
+        "name" : "SentryTestServerTests"
+      }
+    }
+  ],
+  "version" : 1
+}
